requirements.txt
absl-py==1.4.0
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
altgraph==0.17.4
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
blinker==1.7.0
Brotli==1.1.0
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.2.0
colorama==0.4.6
contourpy==1.1.0
cryptography==41.0.5
cycler==0.11.0
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.7
django-allauth==0.58.1
django-ckeditor==6.7.0
django-etc==1.4.0
django-hitcount==1.3.5
django-js-asset==2.1.0
django-mathfilters==1.0.0
django-template-maths==0.2.0
docxcompose==1.4.0
docxtpl==0.16.7
environs==9.5.0
et-xmlfile==1.1.0
exceptiongroup==1.1.2
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.4.0
h11==0.14.0
h2==4.1.0
HandTrackingModule==0.1
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
imap-tools==1.5.0
Jinja2==3.1.2
json-extract==1.1.2
kaitaistruct==0.10
kiwisolver==1.4.4
lxml==4.9.2
magic-filter==1.0.11
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.1
mediapipe==0.10.1
moviepy==1.0.3
multidict==6.0.4
numpy==1.25.0
oauthlib==3.2.2
opencv-contrib-python==4.8.0.74
opencv-python==4.8.0.74
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.1
pandas==2.0.3
pefile==2023.2.7
Pillow==9.5.0
platformdirs==3.11.0
proglog==0.1.10
protobuf==3.20.3
pyasn1==0.5.1
pycparser==2.21
pydivert==2.1.0
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-docx-replace==0.4.4
python-dotenv==1.0.0
python-math==0.0.1
python-telegram-bot==20.4
python3-openid==3.2.0
pytz==2023.3
pywin32-ctypes==0.2.2
requests==2.31.0
requests-oauthlib==1.3.1
scipy==1.11.3
selenium==4.15.2
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.5
sqlparse==0.4.4
threadpoolctl==3.2.0
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.3
virtualenv==20.24.5
wikipedia==1.4.0
wsproto==1.2.0
yarl==1.9.2
zstandard==0.22.0
